Remove PriorityQueue

+++ src/Module.Server/Balancing/PartitioningResult.cs
@@ -11,7 +11,7 @@ public class PartitioningResult<T>
     /// <param name="partition">The partition as a list of arrays of integers. The
     /// integers represent the indices in a number set.</param>
     /// <param name="sizes">The sums of the values in the partition.</param>
-    public PartitioningResult(List<T>[] partition, double[] sizes)
+    public PartitioningResult(List<T>[] partition, float[] sizes)
     {
         Partition = partition;
         Sizes = sizes;
@@ -26,9 +26,9 @@ public class PartitioningResult<T>
     /// <summary>
     /// The sums of the values in the partition.
     /// </summary>
-    public double[] Sizes { get; }
+    public float[] Sizes { get; }
 
-    public void Deconstruct(out List<T>[] partition, out double[] sizes)
+    public void Deconstruct(out List<T>[] partition, out float[] sizes)
     {
         partition = Partition;
         sizes = Sizes;
